+++ firmware/pump_dac.cpp
@@ -37,7 +37,7 @@ void SetPumpCurrentTarget(int32_t microampere)
     pumpDac.SetDuty(volts / VCC_VOLTS);
 }
 
-uint16_t GetPumpOutputDuty()
+float GetPumpOutputDuty()
 {
     return pumpDac.GetLastDuty();
 }
